Main implementation of #148.
TODO:
- restore legend
- improve navigation between mapview <-> wheelview
- bugfixes
