+++ src/app/components/belaqi-map-slider/boundary-canvas.ts
