+++ src/app/components/components.module.ts
@@ -11,6 +11,7 @@ import { AnnualMeanPanelInformationPopupComponent } from './annual-mean-panel/an
 import { AnnualMeanPanelComponent } from './annual-mean-panel/annual-mean-panel.component';
 import { BelaqiChartInformationComponent } from './belaqi-chart/belaqi-chart-information.component';
 import { BelaqiChartComponent } from './belaqi-chart/belaqi-chart.component';
+import { BelaqiMapSliderComponent } from './belaqi-map-slider/belaqi-map-slider.component';
 import { BelaqiUserLocationSliderComponent } from './belaqi-user-location-slider/belaqi-user-location-slider.component';
 import { BelaqiWheelInformationComponent } from './belaqi-wheel/belaqi-wheel-information.component';
 import { BelaqiWheelComponent } from './belaqi-wheel/belaqi-wheel.component';
@@ -59,6 +60,7 @@ const COMPONENTS = [
   BelaqiChartComponent,
   BelaqiChartInformationComponent,
   BelaqiUserLocationSliderComponent,
+  BelaqiMapSliderComponent,
   BelaqiWheelComponent,
   BelaqiWheelInformationComponent,
   CommonSettingsComponent,
